# public/openapi.yml
openapi: "3.0.2"
info:
  title: SQL Query Plugin API
  version: 1.0.0
paths:
...
  /relationships/{relationshipName}:
    get:
      summary: Get relationship information
      operationId: getRelationship
      parameters:
        - in: path
          name: relationshipName
          schema:
            type: string
          required: true
          description: The name of the relationship
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: The name of the relationship
                  models:
                    type: array
                    items:
                      type: string
                      description: The names of the models involved in the relationship
                  joinType:
                    type: string
                    description: The type of join for the relationship
                  condition:
                    type: string
                    description: The condition for the relationship
...
  /functions/{functionName}:
    get:
      summary: Get function details
      operationId: getFunctionDetails
      parameters:
        - in: path
          name: functionName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  parameters:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        type:
                          type: string
                  example:
                    type: array
                    items:
                      type: string
                  description:
                    type: string